CV REPROACH
It is developed with ASP.NET Core MVC5. It is a personal and fully dynamic CV website. There are 2 panels in the project. The default panel includes sections for education, work experience, skills, projects, and a contact form. In the Admin Panel, all this information is dynamically retrieved from the database.
The structures used ın the project: HTML-CSS-Bootstrap-ASP.NET Core-Entity framework Corecrud-SQL Server-Authorize